ASoC: SOF: Implement dynamic pipeline feature
Update the sof_widget_list_setup/free routines to setup
and free connected DAPM widgets when a PCM is opened/closed.
These routines accept a list of connected DAPM widgets as input
and determine the SOF widgets, their corresponding pipeline
widgets and connections between them that need to be setup before
the PCM is triggered.

Also, update the struct snd_sof_route to add a new field to indicate
whether the route has already been setup or not. This will be used
to prevent common routes belonging to connected DAPM widgets in
multiple PCMs from being setup again.

For older firmware that doesn't support the dynamic pipeline
feature, the dynamic_pipeline_widget flag will be set to 0
and the pipelines will never be freed after boot. For newer
firmware, this flag allows for selectively choosing pipelines to be
dynamic/static.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>