ASoC: SOF: add module support for ipc4
The key concept in ipc4 pipeline is module which is
a king of component in current sof fw. There are couples
of modules in fw of ipc4 version and they are indentified
by uuid. Each pipeline will use some instances of these
modules to do audio operations. Each module in a pipeline
has a module id, which is allocated by fw manifest, and
instance id which is allocated by driver. Each module also
has a module config which is initiated by driver.

This patch adds support for module id and instance id and
process modules according to uuid.

Signed-off-by: Rander Wang <rander.wang@intel.com>