ASoC: SOF: move DAI_CONFIG IPC to FE hw_params
Today, we set up all widgets required for all PCM streams
at the time of topology parsing even if they are not
used. An optimization would be to only set up the widgets
required for currently active PCM streams. This would give
the FW the opportunity to power gate unused memory blocks,
thereby saving power.

In preparation for implementing this feature, this patch
changes the hw_params ioctl implementation to determine
the widgets required for a PCM stream by querying
the list of connected DAPM widgets. This list is saved as
part of snd_sof_pcm_stream struct and will be used to setup
the widgets.

Also, because the widgets will only be setup during FE hw_params,
the BE hw_params/free callback functions for ALH and HDA are modified
to simply update the DAI config. The IPC to update the link IDs
in the FW will be sent during FE hw_params. So, for the moment
the only change this patch achieves is to move the DAI_CONFIG IPC
to the sof_widget_list_setup/free() functions that get invoked during
PCM hw_params/hw_free.

These functions will be updated in the following patches to actually
setup/free the widgets themselves for dynamic pipelines.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>