ASoC: SOF: Introduce widget use_count
Add a new field, use_count to struct snd_sof_widget to keep track
of the usage count for each widget. Since widgets can belong to
multiple pipelines, this field will ensure that the widget
is setup only when the first pipeline that needs it is started
and freed when the last pipeline that needs it is stopped. Also,
add a use_count_mutex to protect the access to a widget's use_count.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>