ASoC: SOF: topology: set pipe_widget for all widgets
Introduce a new token in topology to indicate if a pipeline
should be set up dynamically or statically at the time of
topology loading. The parsed token value is stored in the
dynamic_pipeline_widget field of the pipeline widget.

For dynamic pipelines, the pipeline widget(s) corresponding to
each PCM will need to be set up at runtime. Introduce a new field,
pipe_widget in struct snd_sof_widget that saves the pointer to
the pipeline widget that the widget belongs to.

The dynamic_pipeline_widget flag for all widgets in a pipeline is
also updated to match that of the pipeline widget.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>